Add comments to example.anchor-config.yaml and simplified test config files.
