+++ platform/src/main/resources/anchor-config-defaults.yaml
@@ -1,29 +1,76 @@
+# This file configures the anchor platform server.
+
 stellar:
   anchor:
-    # Configure the application from this file after it is loaded in memory
-    # If `config` is in-memory, the yaml file contains all settings for the server.
+    # Set the source of the configuration. The configuration source determines how the anchor platofrm
+    # server fetch the configuration values.
+    #
+    # If `config` is `in-memory`, this yaml file contains all settings for the server.
+    #
+    # `in-memory` is the only `config` value supported in MVP.
+    #
     config: in-memory
+
+    #
+    # Sets how the application configuration values are read.
+    #
+    #  If `type` is `config-spring-property`, the platform server will use the Spring's `@ConfigruationProperties` to
+    #      read populate the configuration values.
+    #
+    #  `config-spring-property` is the only `type` value supported in MVP.
     app-config:
-      type: config-spring-property                # Activate [config-spring-property] module
-      settings: app-config                        # The location of the configuration data
+      type: config-spring-property
+      settings: app-config                        # The location of the configuration data in this yaml file
+
+    #
+    # Sets how the application access the datastore.
+    #
+    # If the `type` is `data-spring-jdbc`, Spring Data JDBC will be used to read the application data.
+    #
+    # `data-spring-jdbc` is the only `type` value supported in MVP.
+    #
     data-access:
-      type: data-spring-jdbc                      # Activate [config-spring-jdbc] module.
-      settings: data-spring-jdbc-sqlite           # The location of the configuration data in this file.
+      type: data-spring-jdbc
+      settings: data-spring-jdbc-sqlite           # The location of the configuration data in this yaml file
+
+
+    #
+    # Sets how the generates the logging information
+    #
+    # If the `type` is `logging-logback`, Logback will be used to generate the logs.
+    #
+    # `logging-logback` is the only `type` value supported in MVP.
+    #
     logging:
       type: logging-logback
-      settings: logging-logback-settings
+      settings: logging-logback-settings          # The location of the configuration data in this yaml file
 
-#
+# ************************************
 # Application settings
-#
+# ************************************
 app-config:
-  # general
+  # shared / general application configurations
   app:
+    # The stellar network passphrase used to access Horizon server.
+    # For Pubnet: use 'Public Global Stellar Network ; September 2015'
+    # For Testnet: use 'Test SDF Network ; September 2015'
     stellarNetworkPassphrase: Test SDF Network ; September 2015
-    hostUrl: http://localhost:8080
-    languages: en
+
+    # The endpoint of the horizon server.
     horizonUrl: https://horizon-testnet.stellar.org
+
+    # The URL where the platform server can be accessed by the client.
+    hostUrl: http://localhost:8080
+
+    # The supported languages defined by RFC4646 (https://datatracker.ietf.org/doc/html/rfc4646)
+    languages: en
+
+    # Location of the assets file in `resources` folder.
+    # TODO: Reading the asset JSON file outside the resource filder should be supported.
     assets: assets-test.json
+
+    # The secret key of JWT encryption
+    # TODO: Move the secret to a separate secret management module.
     jwtSecretKey: secret
 
   # sep-1
@@ -41,35 +88,67 @@ app-config:
   # sep-10
   sep10:
     enabled: true
+
+    # The `home_domain` property of SEP-10. https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0010.md#request
     homeDomain: localhost:8080
+
+    # Set if the client attribution is required
     clientAttributionRequired: false
-    clientAttributionAllowList: lobstr.co,preview.lobstr.co # use this if we want to white list
-    # clientAttributionDenyList: # use this if we want to black list.
+
+    # Set the white list of the client domain. The domains are comma-separated
+    # Although it is allowed to set both the allow-list and the deny-list, it is not helpful if both are set.
+    clientAttributionAllowList: lobstr.co,preview.lobstr.co
+
+    # Set the black list of the client domain.
+    # clientAttributionDenyList:
+
+    # Set the authentication challenge transaction timeout. An expired signed transaction will be rejected.
+    # This is the timeout period the client must finish the authentication process. (ie: sign and respond the challenge
+    # transaction).
     authTimeout: 900
+
+    # Set the timeout of the authenticated JSON Web Token. An expired JWT will be rejected.
+    # This is the timeout period after the client has authenticated.
     jwtTimeout: 86400
+
+    # The private key of the SEP-10 challenge.
+    # We highly recommend that this private key should not be used to sign any transactions to submit to the Stellar
+    # network.
     signingSeed: SAX3AH622R2XT6DXWWSRIDCMMUCCMATBZ5U6XKJWDO7M2EJUBFC3AW5X
 
   # sep-12
   sep12:
     enabled: true
+    # The callback API endpoint.
     customerIntegrationEndpoint: http://localhost:8081
 
   # sep-24
   sep24:
     enabled: true
+    # Set the timeout of the JSON Web Token returned with the embedded interactive url of the SEP-24 process.
+    # An expired JWT will be rejected.
+    #
+    # If the interactive flow needs to access the platform server, the interactive process must finish within
+    # the specified timeout period.
     interactiveJwtExpiration: 3600
+
+    # The interactive URL where the platform server will redirect to start the SEP-24 interactive flow.
     interactiveUrl: http://localhost:8081/sep24/interactive
 
   # sep-31
   sep31:
     enabled: true
+
+    # The callback API endpoint.
     feeIntegrationEndPoint: http://localhost:8081
+
     #
     # paymentType: used to determine how amount_in is calculated from amount in the POST /transaction call
     # Possible values: STRICT_SEND or STRICT_RECEIVE. default=STRICT_SEND
     #     STRICT_SEND: amount_in = amount
     #     STRICT_RECEIVE: amount_in = amount + fee
     paymentType: STRICT_SEND
+
     #
     # depositInfoGeneratorType: used to choose how the SEP-31 deposit information will be generated, which includes the
     # deposit address, memo and memo type.
@@ -81,11 +160,13 @@ app-config:
   # sep-38
   sep38:
     enabled: true
+
+    # The callback API endpoint.
     quoteIntegrationEndPoint: http://localhost:8081
 
   circle:
     circleUrl: https://api-sandbox.circle.com
-    apiKey: secret # circle API key
+    apiKey: QVBJX0tFWTpjOWRjOWU3YjgzNTFhZDIwNWNlODEzYjkyZDEzZTU1YzpkY2U1MzVhMTExYjNjODU1YzQzNjMyZTQ5NWJhNDdlZg== # circle API key
 
   payment-gateway:
     #
