--- platform/src/test/resources/application-integration-test.properties
@@ -1,2 +1,0 @@
-stellar.anchor.config=classpath:test-anchor-config.yaml
-server.servlet.context-path=/
