+++ platform/src/test/resources/test-read.yaml
@@ -1,0 +1,3 @@
+# Test Yaml File
+stellar:
+  test: test-value
